adapt imports for display refactoring
