+++ example/graph_rag/load_entity_resolver.py
@@ -9,7 +9,7 @@ from proscenium.vector_database import add_chunks_to_vector_db
 from proscenium.vector_database import embedding_function
 from proscenium.know import knowledge_graph_client
 from proscenium.vector_database import collection_name
-from proscenium.display import collection_panel
+from proscenium.display.milvus import collection_panel
 
 import example.graph_rag.config as config
